// src/Controlador/types/asignarBodega.ts
 import { DatosBodegas } from "./DatosEnviar";
 export interface datosType {
        datos1: DatosBodegas | null;
        datos2: DatosBodegas | null;
        datos3: DatosBodegas | null;
        datos4: DatosBodegas | null;
        datos5: DatosBodegas | null;
        datos6: DatosBodegas | null;
    };